--- src/components/ErrorPage.module.scss
@@ -1,9 +1,0 @@
-.container {
-  padding: 0.5rem 2rem;
-
-  h1 {
-    font-size: 1.25rem;
-    line-height: 1.75rem;
-    margin-block: 0.5rem;
-  }
-}
