+++ src/env.d.ts
@@ -1,4 +1,4 @@
-/// <reference path="../../.astro/types.d.ts" />
+/// <reference path="../.astro/types.d.ts" />
 
 interface ImportMetaEnv {
   readonly CF_BEACON_TOKEN?: string;
