--- src/pages/history.astro
@@ -1,9 +1,0 @@
----
-import { HistoryView } from "@/components/HistoryView";
-import { SITE_NAME } from "@/const";
-import Layout from "@/layouts/Layout.astro";
----
-
-<Layout title=`History - ${SITE_NAME}`>
-  <HistoryView />
-</Layout>
