--- src/pages/posts.astro
@@ -1,12 +1,0 @@
----
-import { ArticleList } from "@/components/ArticleList";
-import { SITE_NAME, ZENN_USERNAME } from "@/const";
-import Layout from "@/layouts/Layout.astro";
-import { getAllPosts } from "@/lib/article";
-
-const posts = await getAllPosts(ZENN_USERNAME, false);
----
-
-<Layout title=`Posts - ${SITE_NAME}`>
-  <ArticleList title={"Posts"} posts={posts} />
-</Layout>
