--- src/pages/works.astro
@@ -1,9 +1,0 @@
----
-import { WorkList } from "@/components/WorkList";
-import { SITE_NAME } from "@/const";
-import Layout from "@/layouts/Layout.astro";
----
-
-<Layout title=`Works - ${SITE_NAME}`>
-  <WorkList />
-</Layout>
